refactor: 添加接口实现检查并重组 connectors 包
- 将 connector.go 和 router.go 合并到 api/connectors/connector.go
- 为所有连接器实现添加显式接口检查 (var _ Interface = (*Impl)(nil))
  - github.Connector
  - gitlab.GitlabConnector
  - websocket.Connector
  - githubprovider.OAuthProvider
  - githubprovider.installationClientResolver
  - githubprovider.oauthClientResolver
- 更新 gateway.go 使用新的包路径
