+++ backend/internal/infra/providers/github/oauth_provider.go
@@ -13,6 +13,7 @@ import (
 
 	gogithub "github.com/google/go-github/v78/github"
 	"github.com/insmtx/SingerOS/backend/internal/api/dto"
+	"github.com/insmtx/SingerOS/backend/internal/api/auth"
 	"github.com/insmtx/SingerOS/backend/config"
 )
 
@@ -22,6 +23,9 @@ const (
 )
 
 var defaultOAuthScopes = []string{"read:user", "user:email", "repo"}
+
+// 确保 OAuthProvider 实现了 auth.AuthorizationProvider 接口
+var _ auth.AuthorizationProvider = (*OAuthProvider)(nil)
 
 // OAuthProvider 实现 GitHub 用户 OAuth 授权接入。
 type OAuthProvider struct {
